requirements.txt
flask==2.3.3
flask-cors==5.0.1
langchain==0.1.1
langchain-community==0.3.21
langchain-google-genai==0.0.5
google-generativeai==0.3.2
pypdf==5.4.0
chromadb==0.4.24
python-dotenv==1.0.0
pydantic==2.5.2
tqdm==4.66.1
streamlit==1.28.2
requests==2.31.0
pandas==2.1.3
python-dotenv==1.0.0